mod_tool for merging sample tada and storing it separately from mods

+++ tools/src/common/endian.h
@@ -47,7 +47,7 @@ namespace nEndian {
 		);
 	}
 
-	constexpr uint16_t fromBig32(uint16_t uwIn)
+	constexpr uint32_t fromBig32(uint32_t uwIn)
 	{
 		return toBig32(uwIn);
 	}
